requirements.txt
setuptools
wheel
flask
pycrypto

